requirements.txt
wandb
ffmpeg-python
opencv-python
timm
ftfy==6.0.3
tensorboardX==2.4.1
# pycocotools==2.0.4
# pycocoevalcap==1.2
datasets
rouge_score
omegaconf
diffusers[torch]==0.14.0
decord==0.6.0
invisible-watermark
colorlog
deepdish
configobj
json_lines
albumentations==1.1.0
opencv-python
pudb==2019.2
imageio==2.14.1
imageio-ffmpeg==0.4.7
pytorch-lightning
omegaconf==2.1.1
test-tube>=0.7.5
streamlit>=0.73.1
setuptools==59.5.0
pillow==9.0.1
einops==0.4.1
torch-fidelity==0.3.0
transformers==4.27.4
torchmetrics==0.6.0
kornia==0.6
clean-fid
h5py
gradio
pytorch-fid
lpips